finished the entire room refactor!
