+++ src/Creeps/RoleHelpers/ClaimerBodyOptsHelper.ts
@@ -19,7 +19,6 @@ import {
     SpawnHelper,
     SpawnApi,
     UserException,
-    RoomHelper,
     MemoryApi
 } from "Utils/Imports/internals";
