+++ src/Room/Room.Structure.Helper.ts
@@ -1,125 +1,6 @@
-import {
-    MemoryApi,
-    WALL_LIMIT,
-    ERROR_WARN,
-    STIMULATE_FLAG,
-    UserException,
-    TOWER_ALLOWED_TO_REPAIR,
-    CreepAllHelper,
-    TOWER_MIN_DAMAGE_THRESHOLD,
-    SpawnHelper,
-    TOWER_MAX_DAMAGE_THRESHOLD
-} from "Utils/Imports/internals";
+import { UserException, MemoryApi, CreepAllHelper, TOWER_MIN_DAMAGE_THRESHOLD, TOWER_MAX_DAMAGE_THRESHOLD, TOWER_ALLOWED_TO_REPAIR, WALL_LIMIT } from "Utils/Imports/internals";
 
-/*
-    Split ideas:
-
-    Structure -
-        Relates to running structures and related functions for that, runX functions,
-        and any helpers to get a target for a structure
-    State -
-        Relating to the state of the room. Anything that checks what kind of room it is,
-        checks the memory for the room, does high level checks on anything relating to
-        how an aspect of the room is currently sitting
-    Tasks -
-        Things like getting repair targets, ie like things that need to be done in the room, and
-        methods relating to making sure the room carries on normal functions
-    Dependent -
-        Anything relating to depedent rooms. Number of remtoe rooms, number of whatever assiged
-        just anything dealing with an unowned room
-
-*/
-// helper functions for rooms
-export class RoomHelper {
-    /**
-     * check if a specified room is owned by you
-     * @param room the room we want to check
-     */
-    public static isOwnedRoom(room: Room): boolean {
-        if (room.controller !== undefined) {
-            return room.controller.my;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * check if a specified room is an ally room
-     * @param room the room we want to check
-     */
-    public static isAllyRoom(room: Room): boolean {
-        // returns true if a room has one of our names or is reserved by us
-        if (room.controller === undefined) {
-            return false;
-        } else if (
-            room.controller.owner !== undefined &&
-            (room.controller.owner.username === "UhmBrock" || room.controller.owner.username === "jakesboy2")
-        ) {
-            return true;
-        } else if (this.isAllyReserved(room)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Check if a room is reserved by an ally
-     * @param room the room we are checking the reservation for
-     */
-    public static isAllyReserved(room: Room): boolean {
-        if (!room || !room.controller) {
-            return false;
-        }
-        return (
-            room.controller.reservation !== undefined &&
-            room.controller.reservation.username !== undefined &&
-            (room.controller.reservation!.username === "UhmBrock" ||
-                room.controller.reservation!.username === "jakesboy2")
-        );
-    }
-
-    /**
-     * check if a room is a source keeper room
-     * @param room the room we want to check
-     */
-    public static isSourceKeeperRoom(room: Room): boolean {
-        // Contains x pos in [1], y pos in [2]
-        const parsedName: any = /^[WE]([0-9]+)[NS]([0-9]+)$/.exec(room.name);
-        const xOffset = parsedName[1] % 10;
-        const yOffset = parsedName[2] % 10;
-        // If x & y === 5 it's not SK, but both must be between 4 and 6
-        const isSK =
-            !(xOffset === 5 && xOffset === 5) && (xOffset >= 4 && xOffset <= 6) && (yOffset >= 4 && yOffset <= 6);
-        return isSK;
-    }
-
-    /**
-     * check if a room is a highway room
-     * @param room the room we want to check
-     */
-    public static isHighwayRoom(room: Room): boolean {
-        // Contains x pos in [1], y pos in [2]
-        const parsedName: any = /^[WE]([0-9]+)[NS]([0-9]+)$/.exec(room.name);
-        // If x || y is divisible by 10, it's a highway
-        if (parsedName[1] % 10 === 0 || parsedName[2] % 10 === 0) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    /**
-     * check if a room is close enough to send a creep to
-     * @param room the room we want to check
-     */
-    public static inTravelRange(homeRoom: string, targetRoom: string): boolean {
-        const routeArray: Array<{ exit: ExitConstant; room: string }> = Game.map.findRoute(
-            homeRoom,
-            targetRoom
-        ) as Array<{ exit: ExitConstant; room: string }>;
-        return routeArray.length < 20;
-    }
+export class RoomHelper_Structure {
 
     /**
      * check if the object exists within a room
@@ -239,70 +120,6 @@ export class RoomHelper {
      */
     public static excecuteEveryTicks(ticks: number): boolean {
         return Game.time % ticks === 0;
-    }
-
-    /**
-     * check if container mining is active in a room (each source has a container in range)
-     * @param room the room we are checking
-     * @param sources the sources we are checking
-     * @param containers the containers we are checking
-     */
-    public static isContainerMining(
-        room: Room,
-        sources: Array<Source | null>,
-        containers: Array<Structure<StructureConstant> | null>
-    ): boolean {
-        // Loop over sources and make sure theres at least one container in range to it
-        let numMiningContainers: number = 0;
-
-        _.forEach(sources, (source: Source) => {
-            if (_.some(containers, (container: StructureContainer) => source.pos.inRangeTo(container.pos, 2))) {
-                numMiningContainers++;
-            }
-        });
-
-        return numMiningContainers === sources.length;
-    }
-
-    /**
-     * Check and see if an upgrader link exists
-     * @param room the room we are checking for
-     */
-    public static isUpgraderLink(room: Room): boolean {
-        // Throw warning if we do not own this room
-        if (!this.isOwnedRoom(room)) {
-            throw new UserException(
-                "Stimulate flag check on non-owned room",
-                "You attempted to check for a stimulate flag in a room we do not own. Room [" + room.name + "]",
-                ERROR_WARN
-            );
-        }
-
-        return MemoryApi.getUpgraderLink(room) !== null;
-    }
-
-    /**
-     * Check if the stimulate flag is present for a room
-     * @param room the room we are checking for
-     */
-    public static isStimulateRoom(room: Room): boolean {
-        // Throw warning if we do not own this room
-        if (!this.isOwnedRoom(room)) {
-            throw new UserException(
-                "Stimulate flag check on non-owned room",
-                "You attempted to check for a stimulate flag in a room we do not own. Room [" + room.name + "]",
-                ERROR_WARN
-            );
-        }
-
-        const terminal: StructureTerminal | undefined = room.terminal;
-        // Check if we have a stimulate flag with the same room name as this flag
-        return _.some(Memory.flags, (flag: FlagMemory) => {
-            if (flag.flagType === STIMULATE_FLAG) {
-                return Game.flags[flag.flagName].pos.roomName === room.name && terminal !== undefined;
-            }
-            return false;
-        });
     }
 
     /**
@@ -531,222 +348,6 @@ export class RoomHelper {
     }
 
     /**
-     * Returns the number of hostile creeps recorded in the room
-     * @param room The room to check
-     */
-    public static numHostileCreeps(room: Room): number {
-        const hostiles = MemoryApi.getHostileCreeps(room.name);
-        return hostiles.length;
-    }
-    /**
-     * Return the number of remote rooms associated with the given room
-     * @param room
-     */
-    public static numRemoteRooms(room: Room): number {
-        const remoteRooms = MemoryApi.getRemoteRooms(room);
-        return remoteRooms.length;
-    }
-
-    /**
-     * get number of associated claim rooms
-     * @param room
-     */
-    public static numClaimRooms(room: Room): number {
-        const claimRooms = MemoryApi.getClaimRooms(room);
-        return claimRooms.length;
-    }
-
-    /**
-     * get number of associated attack rooms
-     * @param room
-     */
-    public static numAttackRooms(room: Room): number {
-        const attackRooms = MemoryApi.getAttackRooms(room);
-        return attackRooms.length;
-    }
-
-    /**
-     * Returns the number of sources in a room
-     * @param room The room to check
-     */
-    public static numSources(room: Room): number {
-        if (!Memory.rooms[room.name].sources) {
-            return room.find(FIND_SOURCES).length;
-        }
-        return Memory.rooms[room.name].sources.data.length;
-    }
-    /**
-     * Returns the number of sources in all remoteRooms connected to room
-     * @param room The room to check the remoteRooms of
-     */
-    public static numRemoteSources(room: Room): number {
-        // TODO: remove sources and structures from the remote room dependent memory itself
-        const remoteRooms: RemoteRoomMemory[] = Memory.rooms[room.name].remoteRooms!;
-        let numSources: number = 0;
-
-        _.forEach(remoteRooms, (rr: RemoteRoomMemory) => {
-            if (!rr) {
-                return;
-            }
-            // Don't consider these sources valid if the controller is reserved by an enemy, or theres defcon 2 >=
-            if (
-                SpawnHelper.isRemoteRoomEnemyReserved(rr) ||
-                (Memory.rooms[rr.roomName] && Memory.rooms[rr.roomName].defcon >= 2)
-            ) {
-                return;
-            }
-
-            let sourcesInRoom: number = 0;
-            if (
-                Memory.rooms[rr.roomName] &&
-                Memory.rooms[rr.roomName].sources &&
-                Memory.rooms[rr.roomName].sources.data
-            ) {
-                sourcesInRoom = Memory.rooms[rr.roomName].sources.data.length;
-            } else {
-                sourcesInRoom = rr.sources.data;
-            }
-            numSources += sourcesInRoom;
-        });
-        return numSources;
-    }
-
-    /**
-     * get number of remote defenders we need
-     * TODO move to spawn helper
-     * @param room The room to check the dependencies of
-     */
-    public static numRemoteDefenders(room: Room): number {
-        const remoteRooms: RemoteRoomMemory[] = Memory.rooms[room.name].remoteRooms!;
-        let numRemoteDefenders: number = 0;
-
-        _.forEach(remoteRooms, (rr: RemoteRoomMemory) => {
-            if (!rr) {
-                return;
-            }
-
-            // If there are any hostile creeps, add one to remoteDefenderCount
-            // Get hostile creeps in the remote room
-            const defconLevel = Memory.rooms[rr.roomName].defcon;
-            if (defconLevel >= 2) {
-                numRemoteDefenders++;
-            }
-        });
-
-        return numRemoteDefenders;
-    }
-
-    /**
-     * get the number of claim rooms that have not yet been claimed
-     * @param room the room we are checking for
-     */
-    public static numCurrentlyUnclaimedClaimRooms(room: Room): number {
-        const allClaimRooms: Array<ClaimRoomMemory | undefined> = MemoryApi.getClaimRooms(room);
-        const ownedRooms: Room[] = MemoryApi.getOwnedRooms();
-        let sum: number = 0;
-
-        // No existing claim rooms
-        if (allClaimRooms[0] === undefined) {
-            return 0;
-        }
-
-        for (const claimRoom of allClaimRooms) {
-            if (
-                !_.some(ownedRooms, ownedRoom => {
-                    if (claimRoom) {
-                        return room.name === claimRoom!.roomName;
-                    }
-                    return false;
-                })
-            ) {
-                ++sum;
-            }
-        }
-
-        return sum;
-    }
-
-    /**
-     * get the number of domestic defenders by the defcon number
-     * TODO move to spawn helper
-     * @param defcon the defcon level of the room
-     * @param isTowers boolean representing if tower exists in room
-     * @returns the number of defenders to spawn
-     */
-    public static getDomesticDefenderLimitByDefcon(defcon: number, isTowers: boolean): number {
-        switch (defcon) {
-            case 2:
-                return isTowers === true ? 0 : 2;
-            case 3:
-                return isTowers === true ? 0 : 2;
-            case 4:
-                return isTowers === true ? 1 : 2;
-        }
-        return 0;
-    }
-
-    /**
-     * convert a room object to a room position object
-     * TODO move to utils/normalize prolly
-     * @param roomObj the room object we are converting
-     */
-    public static convertRoomObjToRoomPosition(roomObj: RoomObject): RoomPosition | null {
-        if (roomObj.room === undefined) {
-            return null;
-        }
-        const x: number = roomObj.pos.x;
-        const y: number = roomObj.pos.y;
-        const roomName: string = roomObj.room!.name;
-        return new RoomPosition(x, y, roomName);
-    }
-
-    /**
-     * check if the first room is a remote room of the second
-     */
-    public static isRemoteRoomOf(dependentRoomName: string, hostRoomName?: string): boolean {
-        // early returns
-        if (!hostRoomName) {
-            const ownedRooms: Room[] = MemoryApi.getOwnedRooms();
-            for (const room of ownedRooms) {
-                const remoteRooms: RemoteRoomMemory[] = MemoryApi.getRemoteRooms(room);
-                if (_.some(remoteRooms, (rr: RemoteRoomMemory) => rr.roomName === dependentRoomName)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-        if (!Memory.rooms[hostRoomName]) {
-            return false;
-        }
-        if (!Game.rooms[hostRoomName]) {
-            return false;
-        }
-
-        const remoteRooms: RemoteRoomMemory[] = MemoryApi.getRemoteRooms(Game.rooms[hostRoomName]);
-        return _.some(remoteRooms, (rr: RemoteRoomMemory) => rr.roomName === dependentRoomName);
-    }
-
-    /**
-     * verify that the object exists in the game
-     * @param id the id we are checking for
-     * @returns if the object exists
-     */
-    public static verifyObjectByID(id: string): boolean {
-        return Game.getObjectById(id) !== undefined;
-    }
-
-    /**
-     * Check if a room has no reservation on it
-     * @param room the room we are checking
-     */
-    public static isNoReservation(room: Room): boolean {
-        if (room.controller) {
-            return room.controller.reservation === undefined;
-        }
-        return false;
-    }
-
-    /**
      * Gets the total storage level for the room (terminal + storage)
      * @param room the room we are checking in
      * @returns number representing amount of energy in the room's store
@@ -764,5 +365,4 @@ export class RoomHelper {
         }
         return storageLevels;
     }
-
 }
