+++ src/Spawn/CreepLimitHelpers/AdvancedStateCreepLimits.ts
@@ -13,11 +13,12 @@ import {
     ROOM_STATE_ADVANCED,
     ROLE_SCOUT,
     MemoryApi,
-    RoomHelper,
     SpawnHelper,
     SpawnApi,
     Normalize,
-    STORAGE_ADDITIONAL_WORKER_THRESHOLD
+    STORAGE_ADDITIONAL_WORKER_THRESHOLD,
+    RoomHelper_State,
+    RoomHelper_Structure
 } from "Utils/Imports/internals";
 
 export class AdvancedStateCreepLimits implements ICreepSpawnLimits {
@@ -54,7 +55,7 @@ export class AdvancedStateCreepLimits implements ICreepSpawnLimits {
         };
 
         const numLorries: number = SpawnHelper.getLorryLimitForRoom(room, room.memory.roomState!);
-        const numRemoteRooms: number = RoomHelper.numRemoteRooms(room);
+        const numRemoteRooms: number = RoomHelper_State.numRemoteRooms(room);
         const minerLimits: number = MemoryApi.getSources(room.name).length;
         let numWorkers: number = 3 + numRemoteRooms;
 
@@ -88,16 +89,16 @@ export class AdvancedStateCreepLimits implements ICreepSpawnLimits {
             claimer: 0
         };
 
-        const numRemoteRooms: number = RoomHelper.numRemoteRooms(room);
-        const numClaimRooms: number = RoomHelper.numClaimRooms(room);
+        const numRemoteRooms: number = RoomHelper_State.numRemoteRooms(room);
+        const numClaimRooms: number = RoomHelper_State.numClaimRooms(room);
         // If we do not have any remote rooms, return the initial remote limits (Empty)
         if (numRemoteRooms <= 0 && numClaimRooms <= 0) {
             return remoteLimits;
         }
         // Gather the rest of the data only if we have a remote room or a claim room
-        const numRemoteDefenders: number = RoomHelper.numRemoteDefenders(room);
-        const numRemoteSources: number = RoomHelper.numRemoteSources(room);
-        const numCurrentlyUnclaimedClaimRooms: number = RoomHelper.numCurrentlyUnclaimedClaimRooms(room);
+        const numRemoteDefenders: number = RoomHelper_State.numRemoteDefenders(room);
+        const numRemoteSources: number = RoomHelper_State.numRemoteSources(room);
+        const numCurrentlyUnclaimedClaimRooms: number = RoomHelper_State.numCurrentlyUnclaimedClaimRooms(room);
 
         // Generate Limits -----
         remoteLimits[ROLE_REMOTE_MINER] = SpawnHelper.getLimitPerRemoteRoomForRolePerSource(
@@ -124,7 +125,7 @@ export class AdvancedStateCreepLimits implements ICreepSpawnLimits {
      */
     private getNumHarvesters(room: Room): number {
         // [Special Case], if we recovered a room and only have 1 harvester (they would be too small to keep up with room)
-        if (RoomHelper.excecuteEveryTicks(40)) {
+        if (RoomHelper_Structure.excecuteEveryTicks(40)) {
             const harvester: Creep | undefined = _.find(
                 MemoryApi.getMyCreeps(room.name, (c: Creep) => c.memory.role === ROLE_HARVESTER)
             );
