+++ src/Spawn/CreepLimitHelpers/StimulateStateCreepLimits.ts
@@ -11,9 +11,9 @@ import {
     ROLE_CLAIMER,
     ROLE_REMOTE_DEFENDER,
     ROOM_STATE_STIMULATE,
-    RoomHelper,
     SpawnHelper,
-    MemoryApi
+    MemoryApi,
+    RoomHelper_State
 } from "Utils/Imports/internals";
 
 export class StimulateStateCreepLimits implements ICreepSpawnLimits {
@@ -68,16 +68,16 @@ export class StimulateStateCreepLimits implements ICreepSpawnLimits {
             claimer: 0
         };
 
-        const numRemoteRooms: number = RoomHelper.numRemoteRooms(room);
-        const numClaimRooms: number = RoomHelper.numClaimRooms(room);
+        const numRemoteRooms: number = RoomHelper_State.numRemoteRooms(room);
+        const numClaimRooms: number = RoomHelper_State.numClaimRooms(room);
         // If we do not have any remote rooms, return the initial remote limits (Empty)
         if (numRemoteRooms <= 0 && numClaimRooms <= 0) {
             return remoteLimits;
         }
         // Gather the rest of the data only if we have a remote room or a claim room
-        const numRemoteDefenders: number = RoomHelper.numRemoteDefenders(room);
-        const numRemoteSources: number = RoomHelper.numRemoteSources(room);
-        const numCurrentlyUnclaimedClaimRooms: number = RoomHelper.numCurrentlyUnclaimedClaimRooms(room);
+        const numRemoteDefenders: number = RoomHelper_State.numRemoteDefenders(room);
+        const numRemoteSources: number = RoomHelper_State.numRemoteSources(room);
+        const numCurrentlyUnclaimedClaimRooms: number = RoomHelper_State.numCurrentlyUnclaimedClaimRooms(room);
 
         // Generate Limits -----
         remoteLimits[ROLE_REMOTE_MINER] = SpawnHelper.getLimitPerRemoteRoomForRolePerSource(
