+++ src/Spawn/SpawnHelper.ts
@@ -11,7 +11,6 @@ import {
     STANDARD_SQUAD,
     UserException,
     MemoryApi,
-    RoomHelper,
     ZEALOT_SOLO_ARRAY,
     STANDARD_SQUAD_ARRAY,
     STALKER_SOLO_ARRAY,
@@ -25,7 +24,9 @@ import {
     NUM_LIFESPANS_FOR_EXTRA_CREEP,
     MAX_WORKERS_UPGRADER_STATE,
     WALL_LIMIT,
-    SCOUT_SPAWN_TICKS
+    SCOUT_SPAWN_TICKS,
+    RoomHelper_Structure,
+    RoomHelper_State
 } from "Utils/Imports/internals";
 
 /**
@@ -392,7 +393,7 @@ export class SpawnHelper {
         let accessibleTiles: number = 0;
         const roomTerrain: RoomTerrain = new Room.Terrain(room.name);
         _.forEach(sources, (source: Source) => {
-            accessibleTiles += RoomHelper.getNumAccessTilesForTarget(source);
+            accessibleTiles += RoomHelper_Structure.getNumAccessTilesForTarget(source);
         });
         return accessibleTiles;
     }
@@ -546,10 +547,10 @@ export class SpawnHelper {
         if (!room) {
             return true;
         }
-        if (RoomHelper.isNoReservation(room)) {
+        if (RoomHelper_State.isNoReservation(room)) {
             return false;
         }
-        return !RoomHelper.isAllyReserved(room);
+        return !RoomHelper_State.isAllyReserved(room);
     }
 
     /**
