+++ src/Utils/Imports/internals.ts
@@ -15,7 +15,8 @@ export * from "Empire/Empire.Api";
 export * from "Events/Event.Api";
 export * from "Memory/Memory.Api";
 export * from "Pathfinding/Pathfinding.Api";
-export * from "Room/Room.Api";
+export * from "Room/Room.State.Api";
+export * from "Room/Room.Structure.Api";
 export * from "Spawn/Spawn.Api";
 // ----------------------------
 // Export All Helpers
@@ -24,7 +25,8 @@ export * from "Utils/UtilHelpers/ConsoleCommands";
 export * from "Creeps/Creep.All.Helper";
 export * from "Empire/EmpireHelper";
 export * from "Events/EventHelper";
-export * from "Room/RoomHelper";
+export * from "Room/Room.State.Helper";
+export * from "Room/Room.Structure.Helper";
 export * from "Spawn/SpawnHelper";
 export * from "Memory/MemoryHelper";
 export * from "Memory/MemoryHelper_Room";
